# Remove unsafe code from UuidHandler