Ref https://github.com/npgsql/npgsql/pull/2297#issuecomment-460244660 😄