@YohDeadfall 2019-02-04
Backported to 4.0.5 via 12467c18d9a346cf8950b40c871578726f609b83 and 791a3efba2daaed58686ef6def036b5af87f0d26.