@khellang 2019-02-04
> Would you mind to do the same with `NumericHandler` in an additional PR?

I'll take a look 👌