@khellang 2019-02-05
I added a benchmark for UUIDs and ran it before and after this PR. These are the results:

## Before

``` ini

BenchmarkDotNet=v0.11.3, OS=Windows 10.0.15063.1506 (1703/CreatorsUpdate/Redstone2)
Intel Core i7-6820HQ CPU 2.70GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=2648434 Hz, Resolution=377.5816 ns, Timer=TSC
.NET Core SDK=3.0.100-preview-010110
  [Host]     : .NET Core 2.2.1 (CoreCLR 4.6.27207.03, CoreFX 4.6.27207.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.2.1 (CoreCLR 4.6.27207.03, CoreFX 4.6.27207.03), 64bit RyuJIT


```
| Method |                Value |     Mean |    Error |   StdDev |         Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |--------------------- |---------:|---------:|---------:|-------------:|------------:|------------:|------------:|--------------------:|
|   Read | 00000(...)00000 [36] | 63.74 ns | 1.322 ns | 2.985 ns | 15,688,666.5 |           - |           - |           - |                   - |
|  Write | 00000(...)00000 [36] | 95.76 ns | 3.276 ns | 9.608 ns | 10,442,773.9 |      0.0170 |           - |           - |                72 B |


## After

``` ini

BenchmarkDotNet=v0.11.3, OS=Windows 10.0.15063.1506 (1703/CreatorsUpdate/Redstone2)
Intel Core i7-6820HQ CPU 2.70GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=2648434 Hz, Resolution=377.5816 ns, Timer=TSC
.NET Core SDK=3.0.100-preview-010110
  [Host]     : .NET Core 2.2.1 (CoreCLR 4.6.27207.03, CoreFX 4.6.27207.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.2.1 (CoreCLR 4.6.27207.03, CoreFX 4.6.27207.03), 64bit RyuJIT


```
| Method |                Value |     Mean |     Error |   StdDev |         Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |--------------------- |---------:|----------:|---------:|-------------:|------------:|------------:|------------:|--------------------:|
|   Read | 00000(...)00000 [36] | 31.18 ns | 0.9147 ns | 2.683 ns | 32,068,301.3 |           - |           - |           - |                   - |
|  Write | 00000(...)00000 [36] | 67.38 ns | 1.8667 ns | 5.416 ns | 14,840,554.1 |      0.0170 |           - |           - |                72 B |

I wonder why the `Write` benchmark allocates?