@YohDeadfall 2019-02-05
> I wonder why the `Write` benchmark allocates?

Interesting. Will take a look.