@roji 2019-02-28
@YohDeadfall @khellang look at https://github.com/dotnet/corefx/pull/34999 and specifically https://github.com/dotnet/corefx/pull/34999/files#diff-1ec4e2a129b5b04546a3db1aff8a6952R17 which optimizes Guid handling in SqlClient. When targeting .NET Core it simply uses the new constructor accepting a ReadOnlySpan<byte>, and for .NET Standard it uses a byte[] taken from ArrayPool<byte>.Shared. That seems like a better approach than what we're doing, since it doesn't make any assumption as to the binary layout of Guid (and also simplifies the code a little).

The only hitch is that we currently don't target netcoreapp - only net and netstandard - but I'm pretty sure we'll find other reasons to do this.

Make sense to you? If so I can submit a PR.