@khellang 2019-02-28
> That seems like a better approach than what we're doing, since it doesn't make any assumption as to the binary layout of Guid (and also simplifies the code a little).

It really isn't any different than doing `guid.ToByteArray()` it still has all the same assumptions about the binary layout as any other method.