@khellang 2019-02-28
The only reason they're able to use that ctor is because SQL Server stores `Guid` with the same endianness as .NET. That's not the case for Postgres, which uses the RFC4122 (UUID) layout.