@khellang 2019-02-28
If you look at [this code](https://github.com/dotnet/corefx/blob/bf09f74c8fe5b6094771266e6beeb36e7447b252/src/Common/src/CoreLib/System/Guid.cs#L49-L53), you can see that they're merely marshalling the memory if the machine is little endian. That means that if the bytes come from Postgres in big endian format, the `Guid` won't be the same in .NET.