@roji 2019-02-28
You're right, I missed that aspect of it. So we'd have to do big->little endian ourselves outside the constructor, which would be extra processing. Never mind then...