@khellang 2019-02-28
> So we'd have to do big->little endian ourselves outside the constructor, which would be extra processing.

Right, and that extra processing is already handled by `[Read|Write]Int[16|32|64]`.

> That seems like a better approach [...] since it doesn't make any assumption as to the binary layout of `Guid`

I don't think that's something to worry about. It would be a major breaking change if the binary layout of `Guid` would ever change. There's even places in the framework itself where it assumes this layout. See https://github.com/dotnet/corefx/blob/master/src/Common/src/CoreLib/System/Buffers/Text/Utf8Formatter/Utf8Formatter.Guid.cs#L194-L217 as an example.