@khellang 2019-02-28
I still think it might be worth cross-compiling against `netcoreapp3.0`, though. As an example, you could replace the internal implementation of `PGUtil.ReverseEndianness` with a call to `BinaryPrimitives.ReverseEndianness`, which [is now a JIT intrinsic](https://github.com/dotnet/coreclr/pull/18398) that will be swapped (no pun intended) with a `BSWAP` instruction.