@roji 2019-02-28
Yep, that's already somewhere in my mind - although maybe open an issue so we can track it (or just a PR doing it 😃 )?

Unless I'm mistaken, we wouldn't actually need to target `netcoreapp3.0` in order to get the optimized intrinsics-driven implementation, as using the ones in System.Memory should already get us that, no?