@YohDeadfall 2019-02-28
The current implementation of the methods had been taken from System.Memory sources before it was published, and I planned to replace them, but completely forgot about that.